Revert "soc: intel: renamed soc from ace30_ptl to ace30"
This reverts commit a654bfbdfa002ac5ab32d8593d762f5ef3dc4868.

This commit broke Zephyr CI, as both the platform name
and SOC kconfigs are used in the SOF module which would also
need an update. Doing such an update is out of scope for a
hotfix.
