--- boards/intel/adsp/intel_adsp_ace30_defconfig
@@ -1,12 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-CONFIG_MAIN_STACK_SIZE=4096
-
-CONFIG_GEN_ISR_TABLES=y
-CONFIG_GEN_IRQ_VECTOR_TABLE=n
-
-CONFIG_BUILD_OUTPUT_BIN=n
-
-CONFIG_DAI_SSP_HAS_POWER_CONTROL=y
-
-CONFIG_DCACHE_LINE_SIZE=64
