+++ boards/intel/adsp/intel_adsp_ace30_ptl.yaml
@@ -1,4 +1,4 @@
-identifier: intel_adsp/ace30/ptl
+identifier: intel_adsp/ace30_ptl
 name: ACE 3.0 Panther Lake Audio DSP
 type: mcu
 arch: xtensa
