+++ soc/intel/intel_adsp/ace/Kconfig.defconfig.ace30_ptl
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 
-if SOC_INTEL_ACE30
+if SOC_INTEL_ACE30_PTL
 
 config MP_MAX_NUM_CPUS
 	default 5
