+++ soc/intel/intel_adsp/ace/Kconfig.soc
@@ -20,7 +20,7 @@ config SOC_INTEL_ACE20_LNL
 	help
 	  ACE 2.0 Lunar Lake PCH
 
-config SOC_INTEL_ACE30
+config SOC_INTEL_ACE30_PTL
 	bool
 	select SOC_SERIES_INTEL_ADSP_ACE
 	help
@@ -32,9 +32,9 @@ config SOC_SERIES
 config SOC_TOOLCHAIN_NAME
 	default "intel_ace15_mtpm" if SOC_INTEL_ACE15_MTPM
 	default "intel_ace15_mtpm" if SOC_INTEL_ACE20_LNL
-	default "intel_ace30_ptl" if SOC_INTEL_ACE30
+	default "intel_ace30_ptl" if SOC_INTEL_ACE30_PTL
 
 config SOC
 	default "ace15_mtpm" if SOC_INTEL_ACE15_MTPM
 	default "ace20_lnl" if SOC_INTEL_ACE20_LNL
-	default "ace30" if SOC_INTEL_ACE30
+	default "ace30_ptl" if SOC_INTEL_ACE30_PTL
