+++ soc/intel/intel_adsp/ace/include/ace30_ptl/adsp_power.h
