+++ soc/intel/intel_adsp/ace/include/ace30_ptl/dmic_regs_ace3x.h
