+++ soc/intel/intel_adsp/common/clk.c
@@ -96,7 +96,7 @@ void adsp_clock_init(void)
 	} else {
 		platform_lowest_freq_idx = ADSP_CPU_CLOCK_FREQ_IPLL;
 	}
-#if CONFIG_SOC_INTEL_ACE30
+#if CONFIG_SOC_INTEL_ACE30_PTL
 	/* Set the Cardinal clock divider to 18 to get 24.576MHz */
 	ACE_DfPMCCU.dfcrodiv &= ACE_CRODIV_CARCDS_MASK;
 	ACE_DfPMCCU.dfcrodiv |= ACE_CRODIV_CARCDS(0x12);
