+++ soc/intel/intel_adsp/soc.yml
@@ -5,7 +5,7 @@ family:
         socs:
           - name: ace15_mtpm
           - name: ace20_lnl
-          - name: ace30
+          - name: ace30_ptl
       - name: intel_adsp_cavs
         socs:
           - name: cavs25
