+++ tests/kernel/mem_protect/sys_sem/boards/intel_adsp_ace30_ptl.conf
