+++ tests/kernel/mutex/sys_mutex/boards/intel_adsp_ace30_ptl.conf
