+++ tests/kernel/queue/boards/intel_adsp_ace30_ptl.conf
