+++ tests/lib/c_lib/thrd/boards/intel_adsp_ace30_ptl.conf
